# desktop/src/lib/api-spec.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "AI 模拟面试 本机接口",
  "description": "桌面前端与 Python 后端之间的本机通道。仅监听回环地址，需 token 认证。",
  "version": "1"
 },
 "paths": {
...
  "/engine/start": {
   "post": {
    "tags": [
     "engine"
    ],
    "summary": "Engine Start",
    "operationId": "engine_start_engine_start_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StartInterviewBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/engine/stop": {
   "post": {
    "tags": [
     "engine"
    ],
    "summary": "Engine Stop",
    "operationId": "engine_stop_engine_stop_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StopInterviewBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StopResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/engine/mute": {
   "post": {
    "tags": [
     "engine"
    ],
    "summary": "Engine Mute",
    "operationId": "engine_mute_engine_mute_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MuteBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/engine/hint": {
   "post": {
    "tags": [
     "engine"
    ],
    "summary": "Engine Hint",
    "operationId": "engine_hint_engine_hint_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/HintBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/engine/code": {
   "post": {
    "tags": [
     "engine"
    ],
    "summary": "Engine Submit Code",
    "operationId": "engine_submit_code_engine_code_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SubmitCodeBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/coding/challenge": {
   "post": {
    "tags": [
     "coding"
    ],
    "summary": "Coding Challenge",
    "operationId": "coding_challenge_coding_challenge_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ComposeChallengeBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CodingChallenge"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/coding/run": {
   "post": {
    "tags": [
     "coding"
    ],
    "summary": "Coding Run",
    "operationId": "coding_run_coding_run_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RunCodeBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunOutcome"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/coding/judge": {
   "post": {
    "tags": [
     "coding"
    ],
    "summary": "Coding Judge",
    "operationId": "coding_judge_coding_judge_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/JudgeCodeBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JudgeOutcome"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/prepare/session": {
   "post": {
    "tags": [
     "prepare"
    ],
    "summary": "Prepare Build Session",
    "operationId": "prepare_build_session_prepare_session_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BuildSessionBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InterviewState"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/prepare/resume": {
   "post": {
    "tags": [
     "prepare"
    ],
    "summary": "Prepare Ingest Resume",
    "operationId": "prepare_ingest_resume_prepare_resume_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/IngestPathBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StoredResume"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/prepare/job-file": {
   "post": {
    "tags": [
     "prepare"
    ],
    "summary": "Prepare Ingest Job File",
    "operationId": "prepare_ingest_job_file_prepare_job_file_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/IngestPathBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StoredJob"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/prepare/job-text": {
   "post": {
    "tags": [
     "prepare"
    ],
    "summary": "Prepare Ingest Job Text",
    "operationId": "prepare_ingest_job_text_prepare_job_text_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/IngestJobTextBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StoredJob"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/prepare/job-synthesize": {
   "post": {
    "tags": [
     "prepare"
    ],
    "summary": "Prepare Synthesize Job",
    "operationId": "prepare_synthesize_job_prepare_job_synthesize_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SynthesizeJobBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StoredJob"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/prepare/diagnose": {
   "post": {
    "tags": [
     "prepare"
    ],
    "summary": "Prepare Diagnose",
    "operationId": "prepare_diagnose_prepare_diagnose_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DiagnoseBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StoredGap"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/review/generate": {
   "post": {
    "tags": [
     "review"
    ],
    "summary": "Review Generate",
    "operationId": "review_generate_review_generate_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/GenerateReviewBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReviewReport"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/review/{session_id}": {
   "get": {
    "tags": [
     "review"
    ],
    "summary": "Review Load",
    "operationId": "review_load_review__session_id__get",
    "parameters": [
     {
      "name": "session_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Session Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "anyOf": [
          {
           "$ref": "#/components/schemas/ReviewReport"
          },
          {
           "type": "null"
          }
         ],
         "title": "Response Review Load Review  Session Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/sessions": {
   "get": {
    "tags": [
     "sessions"
    ],
    "summary": "Sessions List",
    "operationId": "sessions_list_sessions_get",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/SessionSummary"
         },
         "title": "Response Sessions List Sessions Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError",
    "required": [
     "detail"
    ]
   },
   "HintBody": {
    "properties": {
     "auto": {
      "type": "boolean",
      "title": "Auto",
      "default": false
     }
    },
    "type": "object",
    "title": "HintBody"
   },
...
   "MuteBody": {
    "properties": {
     "muted": {
      "type": "boolean",
      "title": "Muted"
     }
    },
    "type": "object",
    "required": [
     "muted"
    ],
    "title": "MuteBody"
   },
   "Ok": {
    "properties": {
     "ok": {
      "type": "boolean",
      "title": "Ok",
      "default": true
     }
    },
    "type": "object",
    "title": "Ok",
    "required": [
     "ok"
    ]
   },
...
   "StartInterviewBody": {
    "properties": {
     "session_id": {
      "type": "integer",
      "title": "Session Id"
     }
    },
    "type": "object",
    "required": [
     "session_id"
    ],
    "title": "StartInterviewBody",
    "description": "只传 session_id：InterviewState 有 37 个字段且内含整份题库，\n让它在前后端往返一遍纯属浪费，后端自己从库里取更可靠。"
   },
   "StopInterviewBody": {
    "properties": {
     "aborted": {
      "type": "boolean",
      "title": "Aborted",
      "default": false
     }
    },
    "type": "object",
    "title": "StopInterviewBody"
   },
...
  }
 }
}